default: adds a wrapped cross-compiler
Wraps the compiler previously built so that we can use the nix magic.
Only used for busybox as it was the smallest diff.

In disk_img: note that coreutils and sed need to be explicitly used to
be prioritized over the binaries created by busybox.

In headers: Install in $out/include to be picked up by the nix-system.
Appearently, only producing $out/$target/include means it does not get
included in NIX_CFLAGS_COMPILE