boot/rpi: switch tty's in cmdline
Switching order in cmdline gives the serial port the root shell,
which is easier for development. While this this shell still only
open /dev/console instead of a tty, it makes life easier for now.

If /init in your initramfs is a script, the commands executed by it
will inherit the primary console - and only it - as their standard
input/output/error, unless the script explicitly assigns another
device for them. - stackoverflow